- increase in-memory threshold to 256k - use DI more consistently - fix namespaces
